+++ src/rs_driver/driver/decoder/block_iterator.hpp
@@ -138,6 +138,7 @@ public:
     this->tss[blk] = this->tss[blk+1] = tss;
   }
 };
+
 template <typename T_Packet>
 class TwoInOneBlockIterator : public BlockIterator<T_Packet>
 {