+++ src/rs_driver/driver/decoder/decoder_mech.hpp
@@ -50,8 +50,6 @@ struct RSDecoderMechConstParam
   float RY;
   float RZ;
 
-
-
   // firing_ts/chan_ts
   double BLOCK_DURATION;
   double CHAN_TSS[128];