+++ src/rs_driver/driver/input/unix/input_sock_select.hpp
@@ -40,6 +40,7 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <sys/types.h>
 #include <sys/socket.h>
 #include <sys/select.h>
+
 namespace robosense
 {
 namespace lidar